Support kwin_wayland_restarter
We no longer need --exit-with-session as plasma_shutdown explicitly
quits kwin gracefully.

We also don't want to relaunch ksplash if kwin restarts.

We run the whole script as we need to restart ksmserver as xserver will
be restarted.

For other services, if they get restored through crash handlers or this
doesn't really matter. DBus unique-ness will take care of it.