[Notifications] Compress removal of notifications
When removing a notification, it is removed from the model, which will cause the
view to update and move up an older notification, if any. If you now delete a bunch
of notifications in quick succession we will create a dialog every time which then
gets deleted again, causing another notification to move up, etc causing high CPU
spikes and plasma freezes.

This patch compresses removing notifications and tries to announce consecuetive
ranges of notifications to be removed in one go.

BUG: 423594
FIXED-IN: 5.21.0