[plasma-shutdown] Avoid timeout when closing ksmserver
closeSession returns only when the task is complete, a sound API, but we
need to remember to increase the timeout.

If a user has unsaved changes in kate, they could handle that prompt
after an indefinite period of time. We would still want to shutdown
afterwards.

Currently plasma-shutdown would abort if it thought ksmserver timed out.

Not seen in real life, just spotted in code review.