[applets/lock_logout] Add explicit options for logout and reboot
With the UI changes in Kicker and the logout screen over the past years
the concept of "leave" with a default shutodwn action has gotten muddy.

Kicker shows all the options explicitly, the logout applet has got left
behind, and is currently a bit broken. The leave option currently always
shutdown regardless.

It's perfectly valid to want to have a short applet shortcut to log out
sometimes, and shutdown in other times.

This patch adds explicit options for logout and reboot, and clarified
the UI presentation of shutdown to match what it currently actually
does.

BUG: 424426
BUG: 402081