[applets/devicenotifier] Improve default actions for removable devices
Right now for removable devices, the default action is either "Mount" or
"Unmount" (depending on whether you have automount turned on or off).

For the common case where you want to mount a device and then
immediately view its contents in the file manager, this requires a
two-step process: mount it, then expand the expandable list item, then
click on "Show in File Manager.

This commit changes the "Mount" action to "Mount and Open", collapsing
the three-step action into just one.

For people who really do want to just mount a disk without opening it, a
"Mount" action that does not also open the file manager is added to the
expandable section of the list item.

Additionally, because the "Show in File Manager" action that's already
in the expandable section of the list item would be redundant with the
one used as the primary action, it is now only shown for mounted
devices.

This mirrors a similar proposed change in the Plasma Vault applet:
https://invent.kde.org/plasma/plasma-vault/-/merge_requests/8