[applets/lock_logout]: Remove confirmation dialog for suspend/hibernate.
When using the lock/logout applet it feels strange, to get only at two
options a confirmation dialog. Additionally it feels odd to answer for
confirmation when neither the option triggered by Shortcuts nor the
Plasma menu asks for confirmation. Suspend and hibernate are the two
options where no data is lost, there is no need to ask for confirmation.