Port away from KNS3::DownloadDialog
The QWidgets components are about to be deprecated
and the QtQuick dialog has a bunch of visual improvements.

Considering that most KCMs already use the QtQuick components
this MR improves the consistency.

See https://invent.kde.org/frameworks/knewstuff/-/merge_requests/80
for the MR which introduced the non-deprecated dialog class.