[lookandfeel] Avoid rendering invisible contents
An opacity of 0 but still visible still results in nodes in the
scenegraph, which is wasteful.  This is shown in gammaray with some
warnings.

Enabled is also bound to visible as if a text field has focus it still
animates the cursor icon even if inivisble, producing wakeups.

BUG: 347772
FIXED-IN: 5.21


(cherry picked from commit 45e0a722fb85bb5d1ab8bef92080e934254b13aa)