[lookandfeel] Fix wake existing screensaver mode with key presses
In commit 45e0a722fb85bb5d1ab8bef92080e934254b13aa we changed the focus
on the textfield with the intention of fixing the cusor blinking.

This had the unfortunate side effect of breaking waking out of the
screensaver mode with just the keyboard.

That screensaver mode was bound to:
```
property bool blockUI: mainBlock.mainPasswordBox.text.length > 0
```
for that to change, we need to have focus.

Instead this patches cursor blinking more explicity.

Doing it only in the lock screen (as opposed to the toolkit) makes sense
as this only happens due to somewhat unusual usage inside the
lockscreen.

BUG: 431286


(cherry picked from commit 2cbd7e4c70d0592d16cdaf58ed43effe4a716aa4)