[taskmanager] Add new task role `IsHidden`
An iconified task/window is in reality hidden. For example, when
`Show Desktop` action is performed, tasks are 'Minimized'. But when
tasks are minimized, they are 'Hidden'.

This property is essential to detect check if a window is truly
iconified or not when the hidden window previews is set to 'Always'.