[applets/digitalclock] Fix showing agenda view when enabling plugins
The agenda view is only shown when at least one event source is active.
When going from 0 to 1 active sources the view wasn't updated correctly.

This was because the visibility of rightColumn was bound to the
visibility of its child, but since the visibility of the parent affects
the visibility of its childred (https://doc.qt.io/qt-5/qml-qtquick-item.html#visible-prop) that resulted in a broken binding.

Bind to the source property directly to avoid this

BUG: 431433
(cherry picked from commit 56064d80ff2e798460967b28243c8117ee6ebaec)