[kcm/users] Avoid half transactions when user cancels auth dialog
Not all calls to accountsservice require authentication and once we hit one that does and the user cancels the prompt then there might be some fields written already.
This is unexpected for the user since they expect the whole transaction to be cancelled.

To avoid this move the call to SetAccountType to the front since that will always trigger a prompt.

It's arguably a workaround for the lack of a transaction concept in AccountsService, but I don't see a better short-term solution.

BUG: 425036