[applets/digital-clock] Port timezone config UI to Kirigami.BasicListItem
At the time this custom delegate was implemented, Kirigami.BasicListitem
didn't have the ability to easily display custom items inside itself.
Now it does, so we can port to that to save some code and increase
consistency.

The appearance is virtually identical and the UX remains the same with
the exception that you can now select an item by clicking anywhere on
the background, not just on one of the radio buttons.