[kcms/lookandfeel] Make scary warning more explicit
The warning that appears when you check the "Use desktop layout from
theme" checkbox is scary, but not actually scary enough. It doesn't make
clear what exactly will be reset. As a result, the user may
underestimate its destructiveness and mistakenly destroy personal
data--for example, text in a sticky note applet.

Accordingly, let's make the text more explicit so people realize it's
scary.

In the long term, I would like for applying a Global Theme to display a
sheet that tells you what it will do, providing the opportunity to tell
it not to do any of those things that you don't want it to do.

But until we have something like that, let's just make this warning
message more explicit:

BUG: 431691
FIXED-IN: 5.22