Move ksmserver's ksplash notifying to ksmserver
Previously the call to ksplash was being emitted by plasma-session. This
isn't used in the systemd startup so it was introduced in a somewhat
messy script.

This script has an issue:

If ksplash is disabled the bus name won't be registered. Typically this
isn't an issue. However to avoid a race condition ksplash registers
itself as being DBus activatable, so that the dbus server holds
dispatching the message.

This combo means ksmserver our unit is considered as not running until
our script times out. This then defers other things in the chain that
depend on ksmserver.

Moving the call into the relevant piece of code into ksmserver unifies
things into a single path and gives us more control.

BUG: 432364


(cherry picked from commit e7e49916712d2d8efc65d7da1721eb1d31dc65d6)