Handle closeSession being called concurrently
Currently if hypothetically closeSession was called multiple times
before the async operation completed the first caller would never get a
reply and then be left hanging potentially forever.

This is unlikely to happen in real life given only the plasma shutdown
binary should be calling this and that's a unique application.