Enforce SESSION_MANAGER is exported before plasmashell is started
Wants says we should bring ksmserver up, but by default that could be in
any order, we need to define it. kcminit also potentially exports
environment variables.

This was written with the correct order in-mind but the wrong keyword
used.  We don't need "Wants" as plasma-core.target already defines that
we will be loading them along with plasmashell.

It is important that ksmserver is up and has completed exporting the
SESSION_MANAGER env, otherwise plasmashell will not propogate it so
applications it launches introducing an X session regression