[startkde] Parse full return value from ListUnitsByNames
The code was written around the assumption that ListUnitsByNames returns
an empty list if the service cannot be found.

This turns out to not be true. An empty list if returned if a requested
name is malformed, but not if it simply isn't available. To do this we
need to parse the returned structure and read the "loadState" entry.

BUG: 433333