[kcms/icons] Collapse buttons if the row is too wide for the display
Prior to this, when resizing the dialogue to a sufficiently small size
(such as happens on Plasma Mobile on the Pinephone), the row would
overflow, resulting in a distinctly unpleasant look with a cut-off
button. This patch fixes this by using a Kirigami ActionToolBar for
the right-side buttons, allowing them to move into an overflow menu
when there isn't sufficient horizontal space.