[Notifications] Offer "Open Containing Folder" whenever we have a destUrl
The restriction on having an amount of files to be processed was to avoid
showing whatever last file had been copied as destination.

However, we already only fall back to using the "destination" label contents
when there is a single file (cf. `NotificationManager::JobPrivate::destUrl()`),
making it safe to assume destUrl is always a predicatble/useful location.

CCBUG: 408778