Remove "Show KRunner" item from desktop context menu
In the past, this was considered to be a way to make KRunner more
discoverable by users.

However today Kickoff exposes all the same runners as KRunner, so its
full functionality is therefore more discoverable in a different way,
so making the KRunner UI itelf more discoverable is not as important.

Furthermore, this was always a questionable way to increase
discoverability in the first place, because the menu item did not expose
the keyboard shortcut, which is the typical way that people show
KRunner. Also, context menus are generally considered shortcuts for
experts; novice users don't tend to use context menus very much. But
experts are the people more likely to already know about KRunner and its
keyboard shortcut, and would never invoke it using the desktop context
menu.

Removing this menu item de-clutters the desktop context menu a bit, and
we do get complaints that the menu is overwhelming because it has too
many things in it. This is a fairly unimportant item that we can remove
without any real consequences.

Finally, Marco didn't like the change to refer to KRunner using the
actual term "KRunner" in the main UI, so this change removes one of the
two places where that happened. :)