Adapt more KCMs to using Kirigami.ActionToolbar for their footer actions
After getting the Icons KCM set up using an ActionToolbar for the row of
buttons at the bottom of the KCM, it seemed like a good idea to adapt
that method for more of the KCMs. This is the first of those patches,
and it adapts the Colors, Cursor Theme, Desktop Theme, Global Theme, and
Style KCMs to using an ActionToolbar (and consequently allowing them to
scale their UI more correctly).

(cherry picked from commit daf612a464a200c213644e59781280d104cd032b)