save layout when corona startup is completed
When the component is completed the corona startup may not be complete
yet.
This causes a race condition when most of the times when a layout
is created from a lajout.js file, applets geometries don't get correctly
saved.
so reinvoke the save timer when the startup completed signal is emitted
from corona

BUG:433799


(cherry picked from commit 110260424669b7424f93b5678a5a2f4ed87ebf6a)