[Image wallpaper] Make the Positioning combobox as equally wide as the other comboboxes
The wallpaper infrastructure sutures multiple form layouts.

The image wallpaper adds its own Positioning combobox, but the combobox's width does not match that of the two comboboxes right above it.

This patch makes the Positioning combobox as wide as the other two by resuing code already present in ConfigurationContainmentAppearance.qml.