Introduce TriangleMouseFilter
Kickoff has a problem that it has a vertical list of things that act on
hover, but the next step involves moving the mouse to the right where
the target could be at any vertical position.

Moving the mouse in the most direct way involves hovering over other
items.

This can be supressed with a timer, but then we are in a trade off of
false positive activation for some and it being slow and annoying.
Neither is great.

This filter determines movement away from a child list view to
navigation within a child list listview based on mouse direction.