[applets/digital-clock] Fix timezone placeholder message positioning
The positioning here was funky as if it was trying to to do something
clever, but instead all it did was cause the text to overlap with the
first list item entry with certain combinations of font size and window
size. Instead let's just do the typical thing.

(cherry picked from commit 9b448752420a611fe442ae82b48304cec1405876)