Set keyboard focus to the applet when expanding or switching the active applet
Currently, applet can lose keyboard focus if you click on one of the buttons in its header, after which if you collapse and expand the applet again, the focus will remain on this button. This is extremely inconvenient if you prefer to use the keyboard to work with the applet (for example, with the Clipboard applet). This patch set keyboard focus back to the applet after switching the active applet (as well as after collapsing/expanding it).