Revert "[lookandfeel] Fix wake existing screensaver mode with key presses"
This reverts commit 13057013d55ae19e76d29b9edc96510e52da2a7a.

That commit set the cursorVisible: propety to true. However that
property only exists in the QtQuickControls 2 TextField, not the QQC1
version. We ported the lock screen to PlasmaComponents3 (which
internally uses QQC2) after Plasma 5.18, meaning that in the LTS
version, we still have a QQC1 TextField so this property does not exist.
Trying to set it causes a runtime error and breaks the lock screen.

BUG: 435233
FIXED-IN: 5.18.7.1