Ensure that splash screen and logout greeter fill the entire screen
Since both the splash screen and the logout greeter have an exclusive
zone of 0, they may be resized to accommodate for other layer shell
surfaces, for example panels.

We need to set an exclusive zone of -1 to ensure that the splash screen
and the logout greeter fill the entire screen.