Don't assume m_containment is there
The containment property is assigned from qml so it can be bull. Also, if this component is used in a plasmoid that can run also in normal applet mode (i.e. folderview) it will be null in that case.

if we don't have a containment assigned yet, just do nothing in the signal handlers as they are supposed to read and write from the containment config

BUG:436041