Change defaults for color, widget style... to match current Global Theme
This commit changes what counts as the default settings to take into account the default
settings of the current Global Theme.

e.g. if you change the Global Theme to Breeze Dark, when you go to the color KCM, clicking
the "Defaults" button will revert to the Breeze Dark color scheme (because it is the default
color scheme of the active Global Theme), rather than Breeze Light.