unbreak XDG_CONFIG_DIRS and by extension autostart
"If $XDG_CONFIG_DIRS is either not set or empty, a value equal to
/etc/xdg should be used."

If XDG_CONFIG_DIRS was not set prior to startplasma, which it ordinarily
would not be, then the previous code would set the value to "$newdir:"
which was silly and also wrong as per the spec quoted above. Instead
default to /etc/xdg/.

Also camelCase the variables to follow our coding style.