[applets/systemtray] Filter configure action out of overflow menu correctly
This code was trying to filter out the configure action but failing
because it was comparing a QMenuItem to a QAction. We need to compare
the QMenuItem's internal QAction in order for the configure action to
ever match.


(cherry picked from commit 827afc3fc276d0d73b4e4d2f08e26c24a3e6ea71)