PowerDevilRunner: Set screen brightness values
powerdevil expects the brightness to be in absolute values not
percentages.
I know this is a blocking call, but we use QDBusInterfaces and
other blocking calls in here so it's not great but is a reasonable
fix for now.
BUG: 437408


(cherry picked from commit 9dc13079970d28c499c5ca77aa7d9c3f0da4eaa0)