[applets/systemtray] Fix highlight on applet pin
Sometimes parent component on the applet is not fully updated - for
example effectiveStatus is not yet read from the model. In such case it
will have default value leading to undesired effect.

Call all `updateHighlightedItem` in `Qt.callLater` so that it is called
in next event loop. As a bonus it is called only once, which can improve
performance.

BUG: 437105
FIXED-IN: 5.22.0


(cherry picked from commit 07b3b10de7d09b8918bd237cf0eca9d3fe868c64)