Revert "Remove "Show KRunner" item from desktop context menu"
This reverts commit 1cb254416ca7c65b31ba321fc7b34f6c06f844e6.

This change was not implemented in the correct way. The context menu is
configurable, so instead of removing the possibility of the menu item
ever being in the menu, we should have simply removed it from the
default set of menu items.

BUG: 437747
FIXED-IN: 5.22
(cherry picked from commit 740dab284a14edddd74bf14902e57e23734c71d5)