[containmentactions/contextmenu] Remove KRunner action from default menu
This commit implements what was agreed upon in
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/690, but
does it correctly by simply disabling the menu item by default, not
removing it entirely. This ensures that the change is only made for new
installs. Existing users' configs will not be touched (because all
values are serialized to the config file already), and any new users
will be able to re-enable this item if they like it.

CCBUG: 437747