remove delegate recycler from available translations view
the recycler did not really serve a purpose here. the delegates are
fairly cheap and there's a very finite amount of them and the amount
isn't going to be in the thousands either

it did however cause spurious warnings because there is a space of time
where model and index isn't available to the delegate and so trying to
get a role out of the model would cause warnings because model is null
initially.
removing the recycler gets rid of the warnings.

I've also changed parent.width to refer to the view explicitly as parent
isn't necessarily defined right out the gate causing excess warnings.
this is also typically what we do in other places.