[applets/devicenotifier] Don't show empty header most of the time
This header needs to be shown only when not in the System Tray and also
when there are multiple  mounted removable devices. That's the only
circumstance under which it actually has anything in it.

BUG: 438351
FIXED-IN: 5.22.1
(cherry picked from commit 11a48a7a22547f0d2340c7be644f1bb1a72c3d53)