[kcms/kdrb] Drop broken legacy colour sync in rdb
The intention of this code was to sync with a selection of really really
old applications.

Colours wuold be written into xrdb but there was also a hook to store
and load defaults for a specific set of really legacy applications,
initally defined in '98 and not updated much since.

This has become very outdated, such as tell xpdf to use kfmclient which
no longer exists.

This was broken in that it sampled KDEGlobals "wm" config groups, which
isn't where most colours are. It must have been broken for years.