xembed-sni-proxy: Add Menu property
Because we did not have this property the adaptor used a default
constructed empty path. An empty path is invalid though and Qt
would refuse to marshall it.
Unconditionally advertise '/' as the menu path even though we don't have
one. Callers looking for a menu there will receive an error that there's
not the interface that they expected at the path.
BUG:439229
FIXED-IN:5.22.3