repair selection change tracking
forgot to emit the change signal in a number of scenarios. this then
broke actually changing the order and removing items from the selection
because the signal is used to forward the current selection list to the
KConfigSkeleton. by not emitting the signal in all change scenarios we'd
effectively never update the stringlist inside the skeleton causing noop
saves.