Use RowLayout for delegate buttons in the Clipboard applet
All applets for placing a group of buttons use a Layout items (RowLayout or ColumnLayout) that has a non-zero spacing by default, while delegate buttons use a Row that has a zero spacing by default. This MR makes the delegate buttons use a RowLayout, which makes them look similar to the rest of the button groups.