[applets/batterymonitor] Explain the reasoning for perf mode inhibition
It may not be obvious to the user that performance mode is disabled or
degraded as a means of preventing excessive heat generation, so let's be
explicit about that.

Also, let's use less certainty in our description of how the computer
detected that it's sitting on your lap, because this may be inaccurate,
and if we use declarative language, it makes the system seem stupid.
Better to make the system seem wishy-washy than stupid. :)