[applets/systemtray] Use "scale with panel" setting when in tablet mode
This setting is more appropriate when in tablet mode, as it makes the
System Tray icons scale up to fill the panel's height/width. When tablet
mode exits, it returns to the default icon size setting (unless the user
already set to always be in "scale with panel" mode.

CCBUG: 439922