Import KAutostart class from KService
The handling of the autostart is exclusively done in plasma.
Also the method of KAutostart do not work cross platform and most apps
simply use the X-KDE-autostart-condition, which is not affected by this MR.

In case we need to configure cross platform autostart logic the KService class
needs to be rewritten anyways. But there is currently no need or plans for it.
By importing this class in plasma-workspace we are more flexible in any case.

Task: https://phabricator.kde.org/T12176