[applets/mediacontroller] Ensure background image and album art image display at the same time
Background image and album art image sometimes do not display at the same
time because there are two Image instances referring to the same path, but
cache will not work because they are loaded simultaneously. Use ShaderEffect
instead of another Image instance to provide the background to fix the bug.

Use State and PropertyChanges to avoid unnecessary reevaluation of width
and height of albumArtBackground.

BUG: 439753
FIXED-IN: 5.23