Split Interactive console into it's own binary
Plasma had an entire text editor for the script engine in process.

The method for loading this was very convoluted, it loaded a QML file
(in plasma-desktop!) which contained an Item(!) which then loaded a QML
component that showed a widget dialog.

It also means loading a very heavy lib KTextEditor in ShellComponents
which is very heavy. This will get loaded in things like the logout
greeter, which is super slow on the pinephone.

We already had the concept of evalating a script remotely, we can just
hook up the print statement and move the whole executable out.

Technically this is an API break on the plasmashell DBus API, but
pragmatically going from void -> something on the return type isn't
going to break anything.